requirements.txt
openai>=1.40.0
feedparser>=6.0.11
beautifulsoup4>=4.12.3
python-dateutil>=2.9.0
requests>=2.31.0
streamlit>=1.37.0
pydantic>=2.8.2
email-validator>=2.1.1
python-dotenv>=1.0.1
tqdm>=4.66.0
pyyaml>=6.0.0

